Add client implementation, clean up warnings
Added new module, client.[h|c], which implements a websocket client.
Also added test_client.c, which passes all autobahn tests agains a
'fuzzing server'. test_client.c will also soon be used to run
performance/concurrency tests against server.c.

Also tried compiling the project with clang and -Weverything, and fixed
all warnings in non-test code. Most of these were related to
signed/unsigned mixing; darray and protocol now use size_t in more
places where appropriate.

+++ src/config.h
@@ -31,12 +31,13 @@
 #ifndef __CONFIG_H_
 #define __CONFIG_H_
 
+#include <stdint.h>
 #include "endpoint.h"
 
 typedef struct
 {
     char* bindaddr; /* addr to bind to, if NULL, all interfaces */
-    int port; /* port the server will listen on */
+    uint16_t port; /* port the server will listen on */
     endpoint_settings endp_settings; /* endpoint settings */
     int max_clients; /* max clients we allow connected */
 } config_server_options;
